Refactor debegging code
